requirements.txt
python-telegram-bot==21.6
yfinance==0.2.43
python-dotenv==1.0.1
ujson==5.10.0